Updated certificate format
The identity namespace of a certificate is embedded in
it as the Subject Organisation.

CSRs also need to have the namespace in the same location.

Bifrost doesn't have a default namespace. It defaults to uuid.Nil.

Bifrost also reads the namespace directly from the CA cert.
The CA cert must also be a valid
bifrost cert containing the namespace.