Web UI works!
Web UI is able to successfully create private keys and have
them signed by the API.

Issuer CA is now natively multi-tenant. It can issue certificates
for multiple namespaces at once.

The web UI is able to load all of the issuer instance's namespaces.
But the namespace switcher is broken.
So it can only request certs in the first namespace.